+++ Data Storage/Infrastructure/Services/ProjectService.cs
@@ -9,37 +9,55 @@ namespace Infrastructure.Services;
 public class ProjectService(ProjectRepository projectRepository) : IProjectService
 {
 	private readonly ProjectRepository _projectRepository = projectRepository;
-	
-	public async Task CreateProjectAsync(ProjectRegistrationForm form)
+
+	public async Task<bool> CreateProjectAsync(ProjectRegistrationForm form)
 	{
-		var entity = ProjectFactory.Create(form)!;
-		await _projectRepository.AddAsync(entity);
+		var entity = ProjectFactory.Create(form);
+		if (entity == null)
+		{
+			return false;
+		}
+
+		return await _projectRepository.AddAsync(entity);
 	}
 
-	public async Task<IEnumerable<Project>> GetProjectsAsync()
+	public async Task<IEnumerable<Project?>> GetProjectsAsync()
 	{
-		var entity = await _projectRepository.GetAsync();
-		return entity.Select(ProjectFactory.Create)!;
+		var entities = await _projectRepository.GetAsync();
+		return entities?.Select(ProjectFactory.Create)!;
 	}
 
-	public async Task<Project> GetProjectByIdAsync(int id)
+	public async Task<Project?> GetProjectByIdAsync(int id)
 	{
-		var entity = await _projectRepository.GetAsync(x => x.Id == id) ?? throw new Exception("Project not found");
-		return ProjectFactory.Create(entity!)!;
+		var entity = await _projectRepository.GetAsync(x => x.Id == id);
+		if (entity == null)
+		{
+			return null;
+		}
+
+		return ProjectFactory.Create(entity);
 	}
 
-	public async Task UpdateProjectAsync(ProjectUpdateForm form)
+	public async Task<bool> UpdateProjectAsync(ProjectUpdateForm form)
 	{
-		var entity = await _projectRepository.GetAsync(x => x.Id == form.Id) ?? throw new Exception("Project not found");
+		var entity = await _projectRepository.GetAsync(x => x.Id == form.Id);
+		if (entity == null)
+		{
+			return false;
+		}
 
 		ProjectFactory.Update(entity, form);
-
-		await _projectRepository.UpdateAsync(entity);
+		return await _projectRepository.UpdateAsync(entity);
 	}
 
-	public async Task DeleteProjectAsync(int id)
+	public async Task<bool> DeleteProjectAsync(int id)
 	{
-		var entity = await _projectRepository.GetAsync(x => x.Id == id) ?? throw new Exception("Project not found");
-		await _projectRepository.RemoveAsync(entity!);
+		var entity = await _projectRepository.GetAsync(x => x.Id == id);
+		if (entity == null)
+		{
+			return false;
+		}
+
+		return await _projectRepository.RemoveAsync(entity!);
 	}
 }
